requirements.txt
Django==1.4
python-dateutil==2.1
six==1.1.0
wsgiref==0.1.2
